Add protocol dto classes #5 - produce API

+++ src/main/java/com/github/nginate/kafka/protocol/messages/response/MetadataResponse.java
@@ -9,6 +9,10 @@ import lombok.EqualsAndHashCode;
 
 import static com.github.nginate.kafka.protocol.types.TypeName.*;
 
+/**
+ * The response contains metadata for each partition, with partitions grouped together by topic. This metadata refers to
+ * brokers by their broker id. The brokers each have a host and port.
+ */
 @Data
 @ApiKey(ApiKeys.METADATA)
 @EqualsAndHashCode(callSuper = true)
@@ -21,17 +25,17 @@ public class MetadataResponse extends Response {
     @Data
     public static class Broker {
         @Type(INT32)
-        private int nodeId;
+        private Integer nodeId;
         @Type(STRING)
         private String host;
         @Type(INT32)
-        private int port;
+        private Integer port;
     }
 
     @Data
     public static class TopicMetadata {
         @Type(INT16)
-        private int topicErrorCode;
+        private Short topicErrorCode;
         @Type(STRING)
         private String topicName;
         @Type(WRAPPER)
@@ -40,25 +44,25 @@ public class MetadataResponse extends Response {
         @Data
         public static class PartitionMetadata {
             @Type(INT16)
-            private int partitionErrorCode;
+            private Short partitionErrorCode;
             @Type(INT32)
-            private int partitionId;
+            private Integer partitionId;
             /**
              * The node id for the kafka broker currently acting as leader for this partition. If no leader exists
              * because we are in the middle of a leader election this id will be -1.
              */
             @Type(INT32)
-            private int leader;
+            private Integer leader;
             /**
              * The set of alive nodes that currently acts as slaves for the leader for this partition.
              */
             @Type(INT32)
-            private int[] replicas;
+            private Integer[] replicas;
             /**
              * The set subset of the replicas that are "caught up" to the leader
              */
             @Type(INT32)
-            private int[] isr;
+            private Integer[] isr;
         }
     }
 }
